requirements.txt
asgiref==3.6.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
cryptography==39.0.1
defusedxml==0.7.1
diff-match-patch==20200713
distlib==0.3.6
Django==4.1.7
django-allauth==0.52.0
django-import-export==3.0.2
et-xmlfile==1.1.0
filelock==3.9.0
gunicorn==20.1.0
idna==3.4
MarkupPy==1.14
oauthlib==3.2.2
odfpy==1.4.1
openpyxl==3.1.1
Pillow==9.4.0
pipenv==2023.2.18
platformdirs==3.0.0
psycopg==3.1.8
psycopg2-binary==2.9.5
pycparser==2.21
PyJWT==2.6.0
python3-openid==3.2.0
PyYAML==6.0
requests==2.28.2
requests-oauthlib==1.3.1
sqlparse==0.4.3
tablib==3.3.0
typing_extensions==4.5.0
urllib3==1.26.14
virtualenv==20.19.0
virtualenv-clone==0.5.7
whitenoise==6.3.0
xlrd==2.0.1
xlwt==1.3.0
